Fit model

Fits various types of regression models including generalized linear models, regularized models, Bayesian models, and machine learning approaches.

Usage

fit_model(
  formula,
  data,
  method = c("greedy_l0", "glm", "glmnet", "cv.glmnet", "xgb", "susie"),
  family = gaussian,
  alpha = 1,
  max_support_size = NULL,
  ...
)

Arguments

formula

An object of class formula with a symbolic description of the model to be fitted.

data

A data.frame containing the variables in the model.

method

A character string indicating the method to fit the model. This can take any of the following choices:

greedy_l0 - Greedy L0 regression

glm - Generalized Linear Model using stats::glm

glmnet, cv.glmnet - Regularized GLM using glmnet::glmnet

xgb - Gradient Boosting using xgboost

susie - SuSiE regression using susieR package

family

A description of the error distribution and link function. See stats::family for details.

alpha

The elasticnet mixing parameter, between 0 and 1. See glmnet::glmnet for details.

max_support_size

Maximum support size for the greedy-L0 (greedy_l0) backend.

...

Additional parameters passed to the underlying model fitting function.

Value

A list containing two data frames:

metrics - Goodness of fit measures

coefficients - Fitted coefficients

Examples

data("example_matrix", package = "inferCSN")
df <- as.data.frame(example_matrix)
fit_model(g1 ~ ., data = df, method = "greedy_l0")
#> $model
#> $model$algorithm
#> [1] "greedy_l0"
#> 
#> $model$support
#> [1] "g5" "g6"
#> 
#> $model$bic
#> [1] -111.8727
#> 
#> $model$rss
#> [1] 29.79499
#> 
#> $model$n_obs
#> [1] 100
#> 
#> 
#> $metrics
#> # A tibble: 1 × 1
#>   r_squared
#>       <dbl>
#> 1     0.699
#> 
#> $coefficients
#> # A tibble: 2 × 2
#>   variable coefficient
#>   <chr>          <dbl>
#> 1 g5            -0.492
#> 2 g6             0.601
#> 
fit_model(g1 ~ ., data = df, method = "glm")
#> $model
#> 
#> Call:  stats::glm(formula = formula, family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)           g2           g3           g4           g5           g6  
#>     1.53195     -0.28687     -0.10918      0.06866     -0.53336      0.44851  
#> 
#> Degrees of Freedom: 99 Total (i.e. Null);  94 Residual
#> Null Deviance:	    108 
#> Residual Deviance: 31.56 	AIC: 182.5
#> 
#> $metrics
#> # A tibble: 1 × 1
#>   r_squared
#>       <dbl>
#> 1     0.708
#> 
#> $coefficients
#> # A tibble: 6 × 5
#>   variable    coefficient std_err statistic         pval
#>   <chr>             <dbl>   <dbl>     <dbl>        <dbl>
#> 1 (Intercept)      1.53    0.410      3.74  0.000322    
#> 2 g2              -0.287   0.189     -1.52  0.133       
#> 3 g3              -0.109   0.0941    -1.16  0.249       
#> 4 g4               0.0687  0.0984     0.697 0.487       
#> 5 g5              -0.533   0.0860    -6.20  0.0000000148
#> 6 g6               0.449   0.185      2.42  0.0173      
#> 
fit_model(g1 ~ ., data = df, method = "glmnet")
#> $model
#> 
#> Call:  glmnetUtils:::glmnet.formula(formula = formula, data = data,      family = family, alpha = alpha) 
#> 
#>    Df  %Dev  Lambda
#> 1   0  0.00 0.70760
#> 2   1  7.87 0.64470
#> 3   2 14.91 0.58740
#> 4   2 24.25 0.53530
#> 5   2 32.00 0.48770
#> 6   2 38.44 0.44440
#> 7   2 43.78 0.40490
#> 8   2 48.21 0.36890
#> 9   2 51.90 0.33620
#> 10  3 54.96 0.30630
#> 11  3 57.58 0.27910
#> 12  3 59.75 0.25430
#> 13  3 61.55 0.23170
#> 14  3 63.05 0.21110
#> 15  3 64.29 0.19240
#> 16  3 65.32 0.17530
#> 17  3 66.18 0.15970
#> 18  3 66.89 0.14550
#> 19  3 67.48 0.13260
#> 20  3 67.97 0.12080
#> 21  3 68.38 0.11010
#> 22  3 68.72 0.10030
#> 23  3 69.00 0.09139
#> 24  3 69.23 0.08327
#> 25  3 69.42 0.07587
#> 26  3 69.58 0.06913
#> 27  3 69.72 0.06299
#> 28  3 69.83 0.05739
#> 29  3 69.92 0.05230
#> 30  3 69.99 0.04765
#> 31  3 70.06 0.04342
#> 32  3 70.11 0.03956
#> 33  3 70.15 0.03605
#> 34  3 70.19 0.03284
#> 35  3 70.22 0.02993
#> 36  4 70.27 0.02727
#> 37  4 70.33 0.02484
#> 38  4 70.38 0.02264
#> 39  4 70.43 0.02063
#> 40  4 70.46 0.01879
#> 41  4 70.49 0.01712
#> 42  4 70.52 0.01560
#> 43  4 70.54 0.01422
#> 44  4 70.55 0.01295
#> 45  4 70.57 0.01180
#> 46  5 70.58 0.01075
#> 47  5 70.61 0.00980
#> 48  5 70.64 0.00893
#> 49  5 70.67 0.00814
#> 50  5 70.69 0.00741
#> 51  5 70.70 0.00675
#> 52  5 70.72 0.00615
#> 53  5 70.73 0.00561
#> 54  5 70.74 0.00511
#> 55  5 70.75 0.00466
#> 56  5 70.75 0.00424
#> 57  5 70.76 0.00386
#> 58  5 70.76 0.00352
#> 59  5 70.77 0.00321
#> 60  5 70.77 0.00292
#> 61  5 70.77 0.00266
#> 62  5 70.78 0.00243
#> 63  5 70.78 0.00221
#> 64  5 70.78 0.00201
#> 65  5 70.78 0.00184
#> 66  5 70.78 0.00167
#> 67  5 70.78 0.00152
#> 68  5 70.78 0.00139
#> 
#> $metrics
#> # A tibble: 1 × 3
#>   lambda r_squared alpha
#>    <dbl>     <dbl> <dbl>
#> 1  0.133     0.675     1
#> 
#> $coefficients
#> # A tibble: 6 × 2
#>   variable    coefficient
#>   <chr>             <dbl>
#> 1 (Intercept)       1.15 
#> 2 g2               -0.126
#> 3 g3                0    
#> 4 g4                0    
#> 5 g5               -0.405
#> 6 g6                0.421
#> 
fit_model(g1 ~ ., data = df, method = "cv.glmnet")
#> $model
#> 
#> Call:  cv.glmnet.formula(formula = formula, data = data, alpha = alpha,      family = family) 
#> 
#> Measure: GLM Deviance 
#> 
#>      Lambda Index Measure      SE Nonzero
#> min 0.03956    32  0.3611 0.06668       3
#> 1se 0.19236    15  0.4156 0.07009       3
#> 
#> $metrics
#> # A tibble: 1 × 3
#>   lambda r_squared alpha
#>    <dbl>     <dbl> <dbl>
#> 1  0.192     0.643     1
#> 
#> $coefficients
#> # A tibble: 6 × 2
#>   variable    coefficient
#>   <chr>             <dbl>
#> 1 (Intercept)      1.04  
#> 2 g2              -0.0833
#> 3 g3               0     
#> 4 g4               0     
#> 5 g5              -0.352 
#> 6 g6               0.408 
#> 
fit_model(g1 ~ ., data = df, method = "xgb")
#> $model
#> ##### xgb.Booster
#> call:
#>   xgboost::xgb.train(params = params, data = training, nrounds = nrounds, 
#>     verbose = 0)
#> # of features: 6 
#> # of rounds:  1000 
#> 
#> $metrics
#> # A tibble: 1 × 1
#>   r_squared
#>       <dbl>
#> 1     0.993
#> 
#> $coefficients
#> # A tibble: 5 × 4
#>   variable    gain cover frequency
#>   <chr>      <dbl> <dbl>     <dbl>
#> 1 g6       0.424   0.158     0.149
#> 2 g5       0.341   0.241     0.302
#> 3 g2       0.188   0.115     0.153
#> 4 g4       0.0403  0.345     0.258
#> 5 g3       0.00571 0.141     0.136
#> 
fit_model(g1 ~ ., data = df, method = "susie")
#> $model
#> $alpha
#>                g2           g3           g4           g5           g6
#> [1,] 8.018771e-02 1.693743e-18 1.749942e-25 2.420473e-25 9.198123e-01
#> [2,] 3.691107e-17 4.071050e-16 4.081689e-10 1.000000e+00 3.535943e-17
#> [3,] 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01
#> [4,] 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01
#> [5,] 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01 2.000000e-01
#> 
#> $mu
#>               g2         g3          g4         g5         g6
#> [1,] -0.61015564  0.3407639  0.09381846 -0.1047191 0.62335194
#> [2,] -0.09215526 -0.1563413 -0.34096071 -0.5098033 0.09059383
#> [3,]  0.00000000  0.0000000  0.00000000  0.0000000 0.00000000
#> [4,]  0.00000000  0.0000000  0.00000000  0.0000000 0.00000000
#> [5,]  0.00000000  0.0000000  0.00000000  0.0000000 0.00000000
#> 
#> $mu2
#>              g2         g3         g4         g5         g6
#> [1,] 0.37562578 0.11945590 0.01213778 0.01430196 0.39190351
#> [2,] 0.01181474 0.02776475 0.11957635 0.26322152 0.01152939
#> [3,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [4,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> [5,] 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 
#> $Xr
#>   [1]  0.352578809  1.288289861 -1.118083351 -0.641340150 -1.115711955
#>   [6]  0.262637381 -0.911405449 -0.315106994  1.521236616 -0.925156289
#>  [11] -0.809084415  0.624946300 -0.732409773  0.114564169  1.152115887
#>  [16]  0.215972214  1.279389656  0.151442964  0.802960266  0.633500948
#>  [21] -0.695237584  1.301444756  1.415342354 -1.000745356  0.874535576
#>  [26] -0.232588221 -0.965145604  0.353575351  1.650236434 -0.717177373
#>  [31] -0.788582906  0.871798843  0.130887593 -0.982569481  0.347330577
#>  [36]  0.034241205 -1.263439097  1.355308981 -0.129825027 -0.199880519
#>  [41] -0.936107285 -0.768245405 -0.414710977  0.725245728  0.502577567
#>  [46] -0.501515614  0.166094682 -0.205848268  1.287945006  0.733458784
#>  [51] -0.446276678  1.427976266 -0.827371433 -0.097871239 -0.613184140
#>  [56] -0.471466432  1.000273195  0.155943916  0.280260673  0.102882867
#>  [61] -0.720612806 -0.962353159  0.169939953 -0.816001705  1.331101279
#>  [66] -0.741156854  0.140436810  1.271618564 -1.227930855 -0.850465713
#>  [71]  1.759997872  0.060564053 -0.720858493 -0.448337544  0.412590269
#>  [76] -0.888966876 -0.911087581  0.001485133 -1.135035098  0.195556076
#>  [81] -0.951520860 -0.663915990 -0.251078474  1.645849157 -0.479461868
#>  [86]  0.023284933  0.145484548 -0.850549337 -0.807245442 -0.900212240
#>  [91]  1.554242702  0.453850265 -0.916855192  1.559040185  1.289666906
#>  [96]  0.142008496  1.235915159 -0.318513231 -1.253275174 -0.868106279
#> 
#> $KL
#> [1]  3.711683e+00  3.795630e+00 -7.450581e-08 -7.450581e-08 -7.450581e-08
#> 
#> $lbf
#> [1] 5.433341e+01 3.532057e+01 7.450580e-08 7.450580e-08 7.450580e-08
#> 
#> $lbf_variable
#>              g2        g3        g4         g5         g6
#> [1,] 53.4194671 15.023261 -1.062193 -0.7378122 55.8592664
#> [2,] -0.9080137  1.492546 15.310666 36.9300063 -0.9509602
#> [3,]  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> [4,]  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> [5,]  0.0000000  0.000000  0.000000  0.0000000  0.0000000
#> 
#> $sigma2
#> [1] 0.3330964
#> 
#> $V
#> [1] 0.3905982 0.2632215 0.0000000 0.0000000 0.0000000
#> 
#> $pi
#> [1] 0.2 0.2 0.2 0.2 0.2
#> 
#> $null_index
#> [1] 0
#> 
#> $converged
#> [1] TRUE
#> 
#> $elbo
#> [1] -118.79908  -94.53992  -94.43482  -94.43478
#> 
#> $niter
#> [1] 4
#> 
#> $intercept
#> [1] 1.045463
#> 
#> $fitted
#>      E47_205      E31_109      E13_466      E61_516      E36_536      E32_345 
#>  1.158663852  2.094374904 -0.311998307  0.164744894 -0.309626911  1.068722424 
#>      E66_449      E71_389      E50_169      E69_388      E10_449      E30_237 
#> -0.105320405  0.490978050  2.327321660 -0.119071245 -0.002999371  1.431031343 
#>      E20_419      E87_561      E92_676       E3_322      E11_746      E83_494 
#>  0.073675271  0.920649212  1.958200931  1.022057257  2.085474699  0.957528008 
#>      E82_207      E24_210      E40_367      E63_710      E91_663      E99_460 
#>  1.609045310  1.439585991  0.110847460  2.107529799  2.221427397 -0.194660313 
#>       E80_95      E48_222      E39_353      E23_143      E38_107      E15_287 
#>  1.680620620  0.573496822 -0.159060561  1.159660395  2.456321478  0.088907671 
#>      E89_212      E19_118      E93_661      E46_289      E62_173      E90_648 
#>  0.017502138  1.677883887  0.936972637 -0.176484437  1.153415621  0.840326249 
#>      E68_367       E79_39      E59_579      E49_323      E44_284      E95_402 
#> -0.457354053  2.161394025  0.676260017  0.606204525 -0.130022242  0.037839638 
#>      E53_316      E17_651      E22_144      E94_435      E34_575      E86_586 
#>  0.391374067  1.531330772  1.308662610  0.304569429  0.972179726  0.600236775 
#>      E14_633      E45_127      E26_378      E72_751      E88_441      E74_564 
#>  2.094030050  1.539543828  0.359808365  2.234061310 -0.021286389  0.708213805 
#>       E5_271      E27_662      E98_151      E73_590      E51_170       E6_655 
#>  0.192900904  0.334618612  1.806358239  0.962028959  1.086345717  0.908967911 
#>      E60_451      E97_508      E76_643      E28_455      E54_758      E85_281 
#>  0.085472238 -0.156268116  0.976024996 -0.009916661  2.137186323  0.064928190 
#>       E2_567      E52_688       E0_437      E41_278      E65_749      E57_686 
#>  0.946521854  2.077703608 -0.421845811 -0.044380669  2.566082916  0.866649096 
#>      E37_452      E56_190      E43_578      E81_275      E33_466       E8_185 
#>  0.085226550  0.357747500  1.218675313 -0.082881832 -0.105002537  0.807570176 
#>      E18_370      E35_606       E4_485      E42_411      E75_342      E29_707 
#> -0.328950054  1.001641120 -0.145435816  0.142169053  0.555006570  2.451934201 
#>       E9_269      E64_549      E78_537      E58_379      E25_443      E67_352 
#>  0.326623175  0.829369977  0.951569591 -0.044464293 -0.001160399 -0.094127196 
#>      E12_755       E7_196      E16_460       E96_65       E21_62      E77_706 
#>  2.360327745  1.259935309 -0.110770148  2.365125228  2.095751950  0.948093540 
#>      E70_798      E84_444       E1_380      E55_363 
#>  2.042000203  0.487571813 -0.447190130 -0.062021235 
#> 
#> $sets
#> $sets$cs
#> $sets$cs$L2
#> [1] 4
#> 
#> $sets$cs$L1
#> [1] 1 5
#> 
#> 
#> $sets$purity
#>    min.abs.corr mean.abs.corr median.abs.corr
#> L2     1.000000      1.000000        1.000000
#> L1     0.940689      0.940689        0.940689
#> 
#> $sets$cs_index
#> [1] 2 1
#> 
#> $sets$coverage
#> [1] 1 1
#> 
#> $sets$requested_coverage
#> [1] 0.95
#> 
#> 
#> $pip
#>           g2           g3           g4           g5           g6 
#> 8.018771e-02 4.440892e-16 4.081688e-10 1.000000e+00 9.198123e-01 
#> 
#> $X_column_scale_factors
#>        g2        g3        g4        g5        g6 
#> 0.9538743 1.0517131 1.0278276 0.9914195 0.9480014 
#> 
#> attr(,"class")
#> [1] "susie"
#> 
#> $metrics
#> # A tibble: 1 × 1
#>   r_squared
#>       <dbl>
#> 1     0.701
#> 
#> $coefficients
#> # A tibble: 5 × 2
#>   variable coefficient
#>   <chr>          <dbl>
#> 1 g2         -5.13e- 2
#> 2 g3         -6.00e-17
#> 3 g4         -1.35e-10
#> 4 g5         -5.14e- 1
#> 5 g6          6.05e- 1
#>